Fix WS auth bypass, REST timing side-channel, and scheduled ban silent drop

+++ Mods/SMLWebSocket/Source/SMLWebSocket/Private/SMLWebSocketServerRunnable.cpp
@@ -428,7 +428,7 @@ bool FSMLWebSocketServerRunnable::PerformHandshake(FClientState& Client)
         // XOR-fold all bytes from both strings — result is 0 only when every
         // byte matches.  Running the full loop regardless of length ensures the
         // comparison takes the same time whether or not the lengths differ.
-        uint8 Diff = static_cast<uint8>(AuthLen ^ ExpectedLen);
+        uint32 Diff = static_cast<uint32>(AuthLen ^ ExpectedLen);
         const int32 CmpLen = FMath::Min(AuthLen, ExpectedLen);
         for (int32 i = 0; i < CmpLen; ++i)
             Diff |= static_cast<uint8>(AuthUtf8.Get()[i]) ^ static_cast<uint8>(ExpectedUtf8.Get()[i]);
